Patch 10/10 fprinthexadecimal() and fprintdecimal() are using pointer size second arguments.

+++ src/base/output/cl_output_hex.cc
@@ -11,7 +11,7 @@
 
 namespace cln {
 
-void fprinthexadecimal (std::ostream& stream, unsigned long x)
+void fprinthexadecimal (std::ostream& stream, uintptr_t x)
 {
 	#define bufsize 16
 	var char buf[bufsize+1];
@@ -27,13 +27,13 @@ void fprinthexadecimal (std::ostream& stream, unsigned long x)
 	#undef bufsize
 }
 
-void fprinthexadecimal (std::ostream& stream, long x)
+void fprinthexadecimal (std::ostream& stream, intptr_t x)
 {
 	if (x >= 0)
-		fprintdecimal(stream,(unsigned long)x);
+		fprintdecimal(stream,(uintptr_t)x);
 	else {
 		fprintchar(stream,'-');
-		fprintdecimal(stream,(unsigned long)(-1-x)+1);
+		fprintdecimal(stream,(uintptr_t)(-1-x)+1);
 	}
 }
